refactor(test): split pdf static-image tests by preparation normalization overlay and protocol behavior
