+++ src/app/overlays/pdf/tests/static_images/mod.rs
@@ -1,0 +1,7 @@
+use super::super::*;
+use super::helpers::*;
+
+mod normalization;
+mod overlay;
+mod preparation;
+mod protocols;
